agent/requirements.txt
psutil>=5.9.0
GPUtil>=1.4.0
requests>=2.31.0
pyyaml>=6.0
